ENH: Document types for diagrams, static and live.
Also, move `LiveModelDocument` out of model editor.
