+++ packages/frontend/src/model/model_editor.tsx
@@ -1,21 +1,11 @@
-import type { ChangeFn, DocHandle } from "@automerge/automerge-repo";
 import { MultiProvider } from "@solid-primitives/context";
 import { useNavigate, useParams } from "@solidjs/router";
-import {
-    type Accessor,
-    Match,
-    Show,
-    Switch,
-    createMemo,
-    createResource,
-    useContext,
-} from "solid-js";
+import { Match, Show, Switch, createResource, useContext } from "solid-js";
 import invariant from "tiny-invariant";
 
-import type { JsonValue, Permissions } from "catcolab-api";
-import type { Uuid } from "catlog-wasm";
+import type { JsonValue } from "catcolab-api";
 import { newAnalysisDocument } from "../analysis/document";
-import { type ReactiveDoc, RepoContext, RpcContext, getReactiveDoc } from "../api";
+import { RepoContext, RpcContext, getReactiveDoc } from "../api";
 import { IconButton, InlineInput } from "../components";
 import {
     type CellConstructor,
@@ -25,139 +15,35 @@ import {
     newFormalCell,
 } from "../notebook";
 import { BrandedToolbar, HelpButton } from "../page";
-import { type TheoryLibrary, TheoryLibraryContext } from "../stdlib";
+import { TheoryLibraryContext } from "../stdlib";
 import type { Theory } from "../theory";
 import { PermissionsButton } from "../user";
-import { type IndexedMap, indexMap } from "../util/indexing";
 import {
     ModelValidationContext,
     MorphismIndexContext,
     ObjectIndexContext,
     TheoryContext,
 } from "./context";
-import type { ModelDocument } from "./document";
+import { type LiveModelDocument, type ModelDocument, enlivenModelDocument } from "./document";
 import { MorphismCellEditor } from "./morphism_cell_editor";
 import { ObjectCellEditor } from "./object_cell_editor";
 import { TheorySelectorDialog } from "./theory_selector";
 import {
     type ModelJudgment,
-    type ModelValidationResult,
     type MorphismDecl,
     type ObjectDecl,
     newMorphismDecl,
     newObjectDecl,
-    validateModel,
 } from "./types";
 
 import "./model_editor.css";
 
 import ChartNetwork from "lucide-solid/icons/chart-network";
 
-/** A model document "live" for editing.
-
-Contains a reactive model document and an Automerge document handle, plus
-various memos of derived data.
- */
-export type LiveModelDocument = {
-    /** The ref for which this is a live document. */
-    refId: string;
-
-    /** The model document, suitable for use in reactive contexts.
-
-    This data should never be directly mutated. Instead, call `changeDoc` or
-    interact directly with the Automerge document handle.
-     */
-    doc: ModelDocument;
-
-    /** Make a change to the model document. */
-    changeDoc: (f: ChangeFn<ModelDocument>) => void;
-
-    /** The Automerge document handle for the model document. */
-    docHandle: DocHandle<ModelDocument>;
-
-    /** Permissions for the ref retrieved from the backend. */
-    permissions: Permissions;
-
-    /** A memo of the formal content of the model. */
-    formalJudgments: Accessor<Array<ModelJudgment>>;
-
-    /** A memo of the indexed map from object ID to name. */
-    objectIndex: Accessor<IndexedMap<Uuid, string>>;
-
-    /** A memo of the indexed map from morphism ID to name. */
-    morphismIndex: Accessor<IndexedMap<Uuid, string>>;
-
-    /** A memo of the double theory that the model is of, if it is defined. */
-    theory: Accessor<Theory | undefined>;
-
-    /** A memo of the result of validation.*/
-    validationResult: Accessor<ModelValidationResult | undefined>;
-};
-
-export function enlivenModelDocument(
-    refId: string,
-    reactiveDoc: ReactiveDoc<ModelDocument>,
-    theories: TheoryLibrary,
-): LiveModelDocument {
-    const { doc, docHandle, permissions } = reactiveDoc;
-
-    const changeDoc = (f: ChangeFn<ModelDocument>) => docHandle.change(f);
-
-    // Memo-ize the *formal* content of the notebook, since most derived objects
-    // will not depend on the informal (rich-text) content in notebook.
-    const formalJudgments = createMemo<Array<ModelJudgment>>(() => {
-        return doc.notebook.cells
-            .filter((cell) => cell.tag === "formal")
-            .map((cell) => cell.content);
-    }, []);
-
-    const objectIndex = createMemo<IndexedMap<Uuid, string>>(() => {
-        const map = new Map<Uuid, string>();
-        for (const judgment of formalJudgments()) {
-            if (judgment.tag === "object") {
-                map.set(judgment.id, judgment.name);
-            }
-        }
-        return indexMap(map);
-    }, indexMap(new Map()));
-
-    const morphismIndex = createMemo<IndexedMap<Uuid, string>>(() => {
-        const map = new Map<Uuid, string>();
-        for (const judgment of formalJudgments()) {
-            if (judgment.tag === "morphism") {
-                map.set(judgment.id, judgment.name);
-            }
-        }
-        return indexMap(map);
-    }, indexMap(new Map()));
-
-    const theory = createMemo<Theory | undefined>(() => {
-        if (doc.theory !== undefined) return theories.get(doc.theory);
-    });
-
-    const validationResult = createMemo<ModelValidationResult | undefined>(() => {
-        const th = theory();
-        return th ? validateModel(th.theory, formalJudgments()) : undefined;
-    });
-
-    return {
-        refId,
-        doc,
-        changeDoc,
-        docHandle,
-        permissions,
-        formalJudgments,
-        objectIndex,
-        morphismIndex,
-        theory,
-        validationResult,
-    };
-}
-
 export default function ModelPage() {
     const params = useParams();
-    const ref = params.ref;
-    invariant(ref, "Must provide model ref as parameter to model page");
+    const refId = params.ref;
+    invariant(refId, "Must provide model ref as parameter to model page");
 
     const rpc = useContext(RpcContext);
     const repo = useContext(RepoContext);
@@ -165,8 +51,8 @@ export default function ModelPage() {
     invariant(rpc && repo && theories, "Missing context for model page");
 
     const [liveDoc] = createResource<LiveModelDocument>(async () => {
-        const reactiveDoc = await getReactiveDoc<ModelDocument>(rpc, ref, repo);
-        return enlivenModelDocument(ref, reactiveDoc, theories);
+        const reactiveDoc = await getReactiveDoc<ModelDocument>(rpc, repo, refId);
+        return enlivenModelDocument(refId, reactiveDoc, theories);
     });
 
     return <Show when={liveDoc()}>{(liveDoc) => <ModelDocumentEditor liveDoc={liveDoc()} />}</Show>;
